refactor(pr06): rename BaseItemSheet.mts -> ItemSheetDnd35e.mts (G5d)
File-only rename to match class name (class is ItemSheetDnd35e). Also corrects a stray .mts extension in the sibling index.mts. Interface name BaseItemSheetRenderContext intentionally retained per spec (file-path rename only).
